+++ src/main/java/org/betonquest/betonquest/compatibility/Compatibility.java
@@ -29,6 +29,7 @@ import org.betonquest.betonquest.compatibility.quests.QuestsIntegrator;
 import org.betonquest.betonquest.compatibility.redischat.RedisChatIntegrator;
 import org.betonquest.betonquest.compatibility.shopkeepers.ShopkeepersIntegrator;
 import org.betonquest.betonquest.compatibility.skript.SkriptIntegrator;
+import org.betonquest.betonquest.compatibility.traincarts.TrainCartsIntegrator;
 import org.betonquest.betonquest.compatibility.vault.VaultIntegrator;
 import org.betonquest.betonquest.compatibility.worldedit.WorldEditIntegrator;
 import org.betonquest.betonquest.compatibility.worldguard.WorldGuardIntegrator;
@@ -257,6 +258,7 @@ public class Compatibility implements Listener {
         register("HolographicDisplays", HolographicDisplaysIntegrator.class);
         register("fake-block", FakeBlockIntegrator.class);
         register("RedisChat", RedisChatIntegrator.class);
+        register("Train_Carts", TrainCartsIntegrator.class);
     }
 
     private void register(final String name, final Class<? extends Integrator> integrator) {